# Stricter localize typing