--- a/hassio/src/dashboard/hassio-update.ts
+++ b/hassio/src/dashboard/hassio-update.ts
@@ -46,11 +46,9 @@ export class HassioUpdate extends LitElement {
     return html`
       <div class="content">
         <h1>
-          ${this.supervisor.localize(
-            "common.update_available",
-            "count",
-            updatesAvailable
-          )}
+          ${this.supervisor.localize("common.update_available", {
+            count: updatesAvailable,
+          })}
           🎉
         </h1>
         <div class="card-group">