--- a/hassio/src/update-available/update-available-card.ts
+++ b/hassio/src/update-available/update-available-card.ts
@@ -312,7 +312,7 @@ class UpdateAvailableCard extends LitElement {
           "addon.dashboard.not_available_version",
           {
             core_version_installed: this.supervisor.core.version,
-            core_version_needed: addonStoreInfo.homeassistant,
+            core_version_needed: addonStoreInfo.homeassistant!,
           }
         );
       }