--- a/src/auth/ha-auth-flow.ts
+++ b/src/auth/ha-auth-flow.ts
@@ -159,11 +159,9 @@ export class HaAuthFlow extends LitElement {
       case "error":
         return html`
           <ha-alert alert-type="error">
-            ${this.localize(
-              "ui.panel.page-authorize.form.error",
-              "error",
-              this._errorMessage
-            )}
+            ${this.localize("ui.panel.page-authorize.form.error", {
+              error: this._errorMessage,
+            })}
           </ha-alert>
           <div class="action">
             <mwc-button raised @click=${this._startOver}>
@@ -317,15 +315,10 @@ export class HaAuthFlow extends LitElement {
   }
 
   private _computeStepDescription(step: DataEntryFlowStepForm) {
-    const resourceKey =
-      `ui.panel.page-authorize.form.providers.${step.handler[0]}.step.${step.step_id}.description` as const;
-    const args: string[] = [];
-    const placeholders = step.description_placeholders || {};
-    Object.keys(placeholders).forEach((key) => {
-      args.push(key);
-      args.push(placeholders[key]);
-    });
-    return this.localize(resourceKey, ...args);
+    return this.localize(
+      `ui.panel.page-authorize.form.providers.${step.handler[0]}.step.${step.step_id}.description`,
+      step.description_placeholders
+    );
   }
 
   private _computeLabelCallback(step: DataEntryFlowStepForm) {