--- a/src/common/structs/handle-errors.ts
+++ b/src/common/structs/handle-errors.ts
@@ -13,45 +13,33 @@ export const handleStructError = (
   for (const failure of err.failures()) {
     if (failure.value === undefined) {
       errors.push(
-        hass.localize(
-          "ui.errors.config.key_missing",
-          "key",
-          failure.path.join(".")
-        )
+        hass.localize("ui.errors.config.key_missing", {
+          key: failure.path.join("."),
+        })
       );
     } else if (failure.type === "never") {
       warnings.push(
-        hass.localize(
-          "ui.errors.config.key_not_expected",
-          "key",
-          failure.path.join(".")
-        )
+        hass.localize("ui.errors.config.key_not_expected", {
+          key: failure.path.join("."),
+        })
       );
     } else if (failure.type === "union") {
       continue;
     } else if (failure.type === "enums") {
       warnings.push(
-        hass.localize(
-          "ui.errors.config.key_wrong_type",
-          "key",
-          failure.path.join("."),
-          "type_correct",
-          failure.message.replace("Expected ", "").split(", ")[0],
-          "type_wrong",
-          JSON.stringify(failure.value)
-        )
+        hass.localize("ui.errors.config.key_wrong_type", {
+          key: failure.path.join("."),
+          type_correct: failure.message.replace("Expected ", "").split(", ")[0],
+          type_wrong: JSON.stringify(failure.value),
+        })
       );
     } else {
       warnings.push(
-        hass.localize(
-          "ui.errors.config.key_wrong_type",
-          "key",
-          failure.path.join("."),
-          "type_correct",
-          failure.refinement || failure.type,
-          "type_wrong",
-          JSON.stringify(failure.value)
-        )
+        hass.localize("ui.errors.config.key_wrong_type", {
+          key: failure.path.join("."),
+          type_correct: failure.refinement || failure.type,
+          type_wrong: JSON.stringify(failure.value),
+        })
       );
     }
   }