… src/components/ha-area-picker.ts
… src/components/ha-assist-pipeline-picker.ts
… src/components/ha-button-related-filter-menu.ts
… src/components/ha-file-upload.ts
… src/components/media-player/ha-media-player-browse.ts
… src/data/automation_i18n.ts
… src/data/device_automation.ts
… src/data/device_registry.ts
… src/data/logbook.ts
… src/data/script_i18n.ts
… src/dialogs/config-entry-system-options/dialog-config-entry-system-options.ts
… src/dialogs/config-flow/show-dialog-config-flow.ts
… src/dialogs/config-flow/step-flow-create-entry.ts
… src/dialogs/more-info/ha-more-info-info.ts
… src/dialogs/notifications/configurator-notification-item.ts
… src/dialogs/quick-bar/ha-quick-bar.ts
… src/fake_data/provide_hass.ts
… src/layouts/hass-tabs-subpage-data-table.ts
… src/mixins/lit-localize-lite-mixin.ts
… src/panels/config/application_credentials/ha-config-application-credentials.ts
… src/panels/config/areas/ha-config-areas-dashboard.ts
… src/panels/config/automation/action/ha-automation-action-row.ts
… src/panels/config/automation/action/types/ha-automation-action-choose.ts
… src/panels/config/automation/condition/ha-automation-condition-editor.ts
… src/panels/config/automation/ha-automation-editor.ts
… src/panels/config/automation/ha-automation-picker.ts
… src/panels/config/automation/trigger/ha-automation-trigger-row.ts
… src/panels/config/blueprint/dialog-import-blueprint.ts
… src/panels/config/cloud/account/cloud-tts-pref.ts
… src/panels/config/cloud/account/dialog-cloud-tts-try.ts