@piitaya 2023-09-25
Hello 👋

Improving `LocalizeFunc` typings is a good thing. But I will not change the localize implementation because many custom cards relies on this `localize` function because it's part of the `hass` object. The backward compatibility should be added.

Also, what's the motivation to move away from `intl-messageformat`? Only bundle size?