@bramkragten 2023-09-25
Why didn't you use `intl-messageformat-tiny` as a dependency?