@bramkragten 2023-09-25
Did you also do a performance comparison? Is there a test done on functionality and browser compatibility? Seems like a lot of things are just removed from `icu-messageformat-parser`?