@bramkragten 2023-09-25
Can we split this PR, make a PR that only changes the places that call `localize` (something that can be done (and will improve) with the current code.

And then a PR that updates the `localize` function itself.