@KTibow 2023-09-25
Okay. So something like this:
PR 1:
- Use a stricter TS def, but allow the function to still accept other formats
PR 2:
- New localize library
- Has backwards compatibility, clears cache when updated
- Smaller bundle sizes
- Better performance:
| Task | Time with old | Time with new |
|--------|--------|--------|
| Localize a simple string | 380ns | 220ns |
| Localize a string with a variable | 720ns | 560ns |