@steverep 2023-09-25
> Because we need to tweak a number of things, like:
> 1. Stricter TS def
> 2. Fixing it so that it allows more than 1 condition per string
> 3. Letting it return html without having to use unsafeHTML by utilizing the format

Why couldn't that just be handled with a patch applied by `yarn`?  Not having it as a dependency means we don't stay on top of updates, which is a big loss IMO.

Also, just an observation that the "tiny" package hasn't been touched in 10 months, whereas `@formatjs` has fairly frequent maintenance.  Not necessarily an indicator of quality of course or functionality of course, but certainly should be noted.