@piitaya 2023-09-25
Yes, the tiny package has only one maintainer. Move to this code means that we are now in charge of the maintenance of this code. I see the benefit of a smaller bundle but I'm not sure it worth it to move from @formatjs to this.