@steverep 2023-09-25
Agree with @piitaya, and 36K is a pretty small savings on the whole bundle.  Yes, it's a much bigger percentage of the entrypoints, but that can be worked on via better splitting/sharing.