@KTibow 2023-09-25
> we shouldn't do this because not a big improvement in bundle size

We should do this because it makes sense to improve stuff, however how small. + half of my lite login page is just localization.

> use a yarn patch

> maintenance problems

I think the best solution is still to copy the library.
- Yarn patches are hard to edit and break installing with `npm` (I know you shouldn't do that but it's more convenient for dev builds). It's easier to just read a local file.
- We shouldn't worry about the lack of maintenance if we were to use the library or copy the library because there's not much to maintain - it's just some really simple code for parsing messages. I'd go as far as to say if there needs to be maintenance or updates you're doing it wrong.