@KTibow 2023-09-26
Converted this PR to the "PR 1" from earlier - the only change is to the type definition, anywhere that needed to be updated to the strict definition, and some other tiny tweaks to the localize function.