@silamon 2023-09-26
There are also some occurrences in the old javascript files. Quick check seems like they are in developer event pages. You can change them but you don't have to if #17869 gets merged.